+++ CHANGELOG.md
@@ -117,6 +117,8 @@ integrals to their own module
 * Updated interfaces to use both interaction list and 2-zone octrees
 * Added magnetic vector potential calculations for direct solvers
 * Fixed a correctness bug in atan2(), causing significant field errors near element surfaces
+
+### [#25](https://github.com/retkins/oersted/pull/25)
 * Added dipole-order multipole expansion for a and h field (from J) calculations in octree
 * Complete octree refactor; removed 3-zone octree
 * Overhauled the python interface and ffi, dramatically reducing the number of functions 
@@ -125,3 +127,10 @@ and complexity of interfacing between rust and python
 for direct solves, and unified the function signatures
 * Combined all solver settings into a `SolverSettings` class 
 * Unified all Python bindings for field calcs to `a_field()`, `h_field()`, and `b_field()`
+
+### [#26](https://github.com/retkins/oersted/pull/33)
+* Added comprehensive octree benchmarks
+* Replaced 'batch' octree traversal with interaction lists using a sparse storage
+method (3x faster than simple batch traversal)
+* Enable monopole/dipole expansions to be selected by the user
+* Various bugfixes and corrections of minor errors